refactor code and add swagger docs
