--- .nyc_output/98c3bc09-7206-42e9-9d53-a310b0d8d5a5.json
@@ -1,1 +1,0 @@
-{}
