--- .nyc_output/a2ec2ec8-54fb-401a-9b18-7229db3532ff.json
@@ -1,1 +1,0 @@
-{}
