--- .nyc_output/a599833c-7151-4b6e-9252-0e901db60ccd.json
@@ -1,1 +1,0 @@
-{}
