--- .nyc_output/processinfo/98c3bc09-7206-42e9-9d53-a310b0d8d5a5.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":648,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js","prefix","-g"],"execArgv":[],"cwd":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task","time":1603539542631,"ppid":14944,"coverageFilename":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\.nyc_output\\98c3bc09-7206-42e9-9d53-a310b0d8d5a5.json","externalId":"","uuid":"98c3bc09-7206-42e9-9d53-a310b0d8d5a5","files":[]}
