--- .nyc_output/processinfo/a2ec2ec8-54fb-401a-9b18-7229db3532ff.json
@@ -1,1 +1,0 @@
-{"parent":"a599833c-7151-4b6e-9252-0e901db60ccd","pid":8980,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\node_modules\\eslint\\bin\\eslint.js","."],"execArgv":[],"cwd":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task","time":1603539546543,"ppid":6860,"coverageFilename":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\.nyc_output\\a2ec2ec8-54fb-401a-9b18-7229db3532ff.json","externalId":"","uuid":"a2ec2ec8-54fb-401a-9b18-7229db3532ff","files":[]}
