--- .nyc_output/processinfo/a599833c-7151-4b6e-9252-0e901db60ccd.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":3068,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Tim\\AppData\\Roaming\\npm\\node_modules\\npm\\bin\\npm-cli.js","run","test"],"execArgv":[],"cwd":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task","time":1603539544459,"ppid":13728,"coverageFilename":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\.nyc_output\\a599833c-7151-4b6e-9252-0e901db60ccd.json","externalId":"","uuid":"a599833c-7151-4b6e-9252-0e901db60ccd","files":[]}
