--- .nyc_output/processinfo/d8409b06-d816-427b-861f-5b66bd0f6237.json
@@ -1,1 +1,0 @@
-{"parent":"a599833c-7151-4b6e-9252-0e901db60ccd","pid":940,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\node_modules\\mocha\\bin\\mocha"],"execArgv":[],"cwd":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task","time":1603539552780,"ppid":14916,"coverageFilename":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\.nyc_output\\d8409b06-d816-427b-861f-5b66bd0f6237.json","externalId":"","uuid":"d8409b06-d816-427b-861f-5b66bd0f6237","files":["C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\app.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\routes\\authRoute.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\validations\\authValidations.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\controllers\\authController.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\util\\Helpers.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\routes\\objectPatchRoute.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\validations\\objectPatchValidation.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\middlewares\\isAuthenticated.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\controllers\\objectPatchController.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\routes\\resizeImageRoute.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\validations\\resizeImageValidations.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\controllers\\resizeImageController.js"]}
