+++ .nyc_output/processinfo/e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":16852,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\node_modules\\mocha\\bin\\mocha","--timeout","15000"],"execArgv":[],"cwd":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task","time":1603577981127,"ppid":17216,"coverageFilename":"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\.nyc_output\\e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c.json","externalId":"","uuid":"e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c","files":["C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\app.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\routes\\authRoute.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\validations\\authValidations.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\controllers\\authController.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\utils\\Helpers.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\routes\\objectPatchRoute.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\middlewares\\isAuthenticated.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\controllers\\objectPatchController.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\routes\\resizeImageRoute.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\validations\\resizeImageValidations.js","C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\controllers\\resizeImageController.js"]}
