+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"98c3bc09-7206-42e9-9d53-a310b0d8d5a5":{"parent":null,"children":[]},"a2ec2ec8-54fb-401a-9b18-7229db3532ff":{"parent":"a599833c-7151-4b6e-9252-0e901db60ccd","children":[]},"a599833c-7151-4b6e-9252-0e901db60ccd":{"parent":null,"children":["a2ec2ec8-54fb-401a-9b18-7229db3532ff","d8409b06-d816-427b-861f-5b66bd0f6237"]},"d8409b06-d816-427b-861f-5b66bd0f6237":{"parent":"a599833c-7151-4b6e-9252-0e901db60ccd","children":[]}},"files":{"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\app.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\routes\\authRoute.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\validations\\authValidations.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\controllers\\authController.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\util\\Helpers.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\routes\\objectPatchRoute.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\validations\\objectPatchValidation.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\middlewares\\isAuthenticated.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\controllers\\objectPatchController.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\routes\\resizeImageRoute.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\validations\\resizeImageValidations.js":["d8409b06-d816-427b-861f-5b66bd0f6237"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\controllers\\resizeImageController.js":["d8409b06-d816-427b-861f-5b66bd0f6237"]},"externalIds":{}}
+{"processes":{"e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c":{"parent":null,"children":[]}},"files":{"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\app.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\routes\\authRoute.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\validations\\authValidations.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\controllers\\authController.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\utils\\Helpers.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\routes\\objectPatchRoute.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\middlewares\\isAuthenticated.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\controllers\\objectPatchController.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\routes\\resizeImageRoute.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\validations\\resizeImageValidations.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"],"C:\\Users\\Tim\\Desktop\\projects\\challenges\\hackerbay_interview\\backend_task\\src\\controllers\\resizeImageController.js":["e5bb49db-5a19-4cf9-87c9-c31c13d8ab3c"]},"externalIds":{}}
