+++ appLogs.js
@@ -1,0 +1,8 @@
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:02.971Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:07.915Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:12.673Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:18.088Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:22.699Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:26.714Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:31.418Z"}
+{"message":"App listening at http://localhost:5000","level":"info","timestamp":"2020-10-24T22:20:36.625Z"}
