+++ bin/www
@@ -1,5 +1,6 @@
 const http = require('http');
 const debug = require('debug')('api:server');
+const logger = require('../logger');
 const app = require('../app');
 
 const server = http.createServer(app);
@@ -52,6 +53,6 @@ const onListening = () => {
 
 /** Listen on provided port, on all network interfaces. **/
 
-server.listen(port, () => console.log(`App listening at http://localhost:${port}`))
+server.listen(port, () => logger.info(`App listening at http://localhost:${port}`))
 server.on('error', onError);
 server.on('listening', onListening);
