+++ logger.js
@@ -1,0 +1,15 @@
+const {
+  createLogger, transports, format
+} = require('winston');
+
+const logger = createLogger({
+  transports: [
+    new transports.File({
+      filename: 'appLogs.js',
+      level: 'info',
+      format: format.combine(format.timestamp(), format.json())
+    })
+  ]
+});
+
+module.exports = logger;
