+++ src/routes/resizeImageRoute.js
