+++ src/validations/resizeImageValidations.js
